minor fixes to some malloc sizeof typing to get rid of some analyzer warnings
git-svn-id: svn://orca.physics.unc.edu/Orca@7352 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Objects/Hardware/Camac/C111C/ORC111CModel.m
@@ -503,7 +503,7 @@ void IRQHandler(short crate_id, short irq_type, unsigned int irq_data,unsigned l
 	blk_info.totsize = numWords;  
     blk_info.ascii_transf = 0;	
 	blk_info.timeout = 0;
-	unsigned int* buffer = (unsigned int*)malloc(numWords*sizeof(int));
+	unsigned int* buffer = (unsigned int*)malloc(numWords*sizeof(unsigned int));
 	if(trackTransactions)[transactionTimer reset];
 	int i;
 	if(f < 16){
@@ -547,7 +547,7 @@ void IRQHandler(short crate_id, short irq_type, unsigned int irq_data,unsigned l
 	blk_info.blksize = 24; //16 bit word size  
 	blk_info.totsize = numWords;  	
 	blk_info.timeout = 0;
-	unsigned int* buffer = (unsigned int*)malloc(numWords*sizeof(long));
+	unsigned int* buffer = (unsigned int*)malloc(numWords*sizeof(unsigned int));
 	if(trackTransactions)[transactionTimer reset];
 	int i;
 	if(f < 16){
